MovieVerse-Mobile/platforms/android/app/src/main/assets/www/MovieVerse-Backend/services/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
faker==24.8.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.7.2
redis==5.0.4
kafka-python==2.0.2
pika==1.3.2
opensearch-py==2.4.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.0
prometheus-client==0.20.0
lxml==5.2.2
